Switch around tests for better diffs
