+++ packages/tests/server/react/polymorphism.legacy.test.tsx
@@ -13,16 +13,12 @@ import {
   QueryClient,
   QueryClientProvider,
   useMutation,
-  useQuery,
 } from '@tanstack/react-query';
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { getUntypedClient } from '@trpc/client';
-import {
-  createTRPCContext,
-  TRPCOptionsProxy,
-} from '@trpc/tanstack-react-query';
-// import type { InferQueryLikeData } from '@trpc/react-query/shared';
+import { createTRPCReact } from '@trpc/react-query';
+import type { InferQueryLikeData } from '@trpc/react-query/shared';
 import { konn } from 'konn';
 import type { ReactNode } from 'react';
 import React, { useState } from 'react';
@@ -34,11 +30,11 @@ import { t } from './polymorphism.common';
  * This also exports types generated by RouterLike and UtilsLike to describe
  * interfaces which concrete router instances are compatible with
  */
-import * as Factory from './polymorphism.query-options.factory';
+import * as Factory from './polymorphism.legacy.factory';
 /**
  * We also define a factory which extends from the basic Factory with an entity sub-type and extra procedure
  */
-import * as SubTypedFactory from './polymorphism.query-options.subtyped-factory';
+import * as SubTypedFactory from './polymorphism.legacy.subtyped-factory';
 
 /**
  * The tRPC backend is defined here
@@ -110,24 +106,25 @@ describe('polymorphism', () => {
       } = createTRPCApi();
 
       const opts = routerToServerAndClientNew(appRouter);
-
-      const { TRPCProvider, useTRPC } = createTRPCContext<typeof appRouter>();
+      const trpc = createTRPCReact<typeof appRouter>();
 
       const queryClient = new QueryClient();
 
       function App(props: { children: ReactNode }) {
         return (
-          <TRPCProvider queryClient={queryClient} trpcClient={opts.client}>
+          <trpc.Provider
+            {...{ queryClient, client: getUntypedClient(opts.client) }}
+          >
             <QueryClientProvider client={queryClient}>
               {props.children}
             </QueryClientProvider>
-          </TRPCProvider>
+          </trpc.Provider>
         );
       }
       return {
         ...opts,
         App,
-        useTRPC,
+        trpc,
         IssueExportsProvider,
         DiscussionExportsProvider,
         PullRequestExportsProvider,
@@ -140,26 +137,25 @@ describe('polymorphism', () => {
 
   describe('simple factory', () => {
     test('can use a simple factory router with an abstract interface', async () => {
-      const { useTRPC } = ctx;
+      const { trpc } = ctx;
 
       /**
        * Can now define page components which re-use functionality from components,
        * and pass the specific backend functionality which is needed need
        */
       function IssuesExportPage() {
-        const trpc = useTRPC();
-        // const utils = trpc.useUtils();
+        const utils = trpc.useUtils();
 
         const [currentExport, setCurrentExport] = useState<number | null>(null);
 
         const invalidate = useMutation({
-          // mutationFn: () => utils.invalidate(),
+          mutationFn: () => utils.invalidate(),
         });
         return (
           <>
             <StartExportButton
               route={trpc.github.issues.export}
-              // utils={utils.github.issues.export}
+              utils={utils.github.issues.export}
               onExportStarted={setCurrentExport}
             />
 
@@ -206,22 +202,21 @@ describe('polymorphism', () => {
     });
 
     test('can use the abstract interface with a factory instance which has been merged with some extra procedures', async () => {
-      const { useTRPC } = ctx;
+      const { trpc } = ctx;
 
       function DiscussionsExportPage() {
-        const trpc = useTRPC();
-        // const utils = trpc.useUtils();
+        const utils = trpc.useUtils();
 
         const [currentExport, setCurrentExport] = useState<number | null>(null);
         const invalidate = useMutation({
-          // mutationFn: () => utils.invalidate(),
+          mutationFn: () => utils.invalidate(),
         });
 
         return (
           <>
             <StartExportButton
               route={trpc.github.discussions.export}
-              // utils={utils.github.discussions.export}
+              utils={utils.github.discussions.export}
               onExportStarted={setCurrentExport}
             />
 
@@ -270,19 +265,18 @@ describe('polymorphism', () => {
 
   describe('sub-typed factory', () => {
     test('can use a sub-typed factory router with the interfaces from the supertype', async () => {
-      const { useTRPC } = ctx;
+      const { trpc } = ctx;
 
       /**
        * Can define page components which operate over interfaces generated by a super-typed router,
        * but also extend types and functionality
        */
       function PullRequestsExportPage() {
-        const trpc = useTRPC();
-        // const utils = trpc.useUtils();
+        const utils = trpc.useUtils();
 
         const [currentExport, setCurrentExport] = useState<number | null>(null);
         const invalidate = useMutation({
-          // mutationFn: () => utils.invalidate(),
+          mutationFn: () => utils.invalidate(),
         });
 
         return (
@@ -290,7 +284,7 @@ describe('polymorphism', () => {
             {/* Some components may still need to be bespoke... */}
             <SubTypedStartExportButton
               route={trpc.github.pullRequests.export}
-              // utils={utils.github.pullRequests.export}
+              utils={utils.github.pullRequests.export}
               onExportStarted={setCurrentExport}
             />
 
@@ -301,7 +295,7 @@ describe('polymorphism', () => {
 
             <RemoveExportButton
               remove={trpc.github.pullRequests.export.delete}
-              // utils={utils.github.pullRequests.export}
+              utils={utils.github.pullRequests.export}
               exportId={currentExport}
             />
 
@@ -358,18 +352,16 @@ describe('polymorphism', () => {
 
 function StartExportButton(props: {
   route: Factory.ExportRouteLike;
-  // utils: Factory.ExportUtilsLike;
+  utils: Factory.ExportUtilsLike;
   onExportStarted: (id: number) => void;
 }) {
-  const exportStarter = useMutation(
-    props.route.start.mutationOptions({
-      async onSuccess(data) {
-        props.onExportStarted(data.id);
+  const exportStarter = props.route.start.useMutation({
+    async onSuccess(data) {
+      props.onExportStarted(data.id);
 
-        // await props.utils.invalidate();
-      },
-    }),
-  );
+      await props.utils.invalidate();
+    },
+  });
 
   return (
     <button
@@ -389,15 +381,13 @@ function StartExportButton(props: {
 function RemoveExportButton(props: {
   exportId: number | null;
   remove: SubTypedFactory.ExportSubTypedRouteLike['delete'];
-  // utils: SubTypedFactory.ExportSubTypesUtilsLike;
+  utils: SubTypedFactory.ExportSubTypesUtilsLike;
 }) {
-  const exportDeleter = useMutation(
-    props.remove.mutationOptions({
-      async onSuccess() {
-        // await props.utils.invalidate();
-      },
-    }),
-  );
+  const exportDeleter = props.remove.useMutation({
+    async onSuccess() {
+      await props.utils.invalidate();
+    },
+  });
 
   const id = props.exportId;
   if (!id) {
@@ -420,19 +410,17 @@ function RemoveExportButton(props: {
 
 type SubTypedStartExportButtonProps = {
   route: SubTypedFactory.ExportSubTypedRouteLike;
-  // utils: SubTypedFactory.ExportSubTypesUtilsLike;
+  utils: SubTypedFactory.ExportSubTypesUtilsLike;
   onExportStarted: (id: number) => void;
 };
 function SubTypedStartExportButton(props: SubTypedStartExportButtonProps) {
-  const exportStarter = useMutation(
-    props.route.start.mutationOptions({
-      onSuccess(data) {
-        props.onExportStarted(data.id);
+  const exportStarter = props.route.start.useMutation({
+    onSuccess(data) {
+      props.onExportStarted(data.id);
 
-        // props.utils.invalidate();
-      },
-    }),
-  );
+      props.utils.invalidate();
+    },
+  });
 
   return (
     <button
@@ -469,14 +457,12 @@ function ExportStatus<
   TStatus extends Factory.ExportRouteLike['status'],
 >(props: {
   status: TStatus;
-  renderAdditionalFields?: (data: TStatus['_output']) => ReactNode;
+  renderAdditionalFields?: (data: InferQueryLikeData<TStatus>) => ReactNode;
   currentExport: number | null;
 }) {
-  const exportStatus = useQuery(
-    props.status.queryOptions(
-      { id: props.currentExport ?? -1 },
-      { enabled: props.currentExport !== null },
-    ),
+  const exportStatus = props.status.useQuery(
+    { id: props.currentExport ?? -1 },
+    { enabled: props.currentExport !== null },
   );
 
   if (!exportStatus.data) {
@@ -487,13 +473,13 @@ function ExportStatus<
     <p>
       Last Export: `{exportStatus.data?.name}` (
       {exportStatus.data.downloadUri ? 'Ready!' : 'Working'})
-      {props.renderAdditionalFields?.(exportStatus.data)}
+      {props.renderAdditionalFields?.(exportStatus.data as unknown as any)}
     </p>
   );
 }
 
 function ExportsList(props: { list: Factory.ExportRouteLike['list'] }) {
-  const exportsList = useQuery(props.list.queryOptions());
+  const exportsList = props.list.useQuery();
 
   return (
     <>
