+++ packages/tests/server/react/polymorphism.test.tsx
@@ -13,12 +13,16 @@ import {
   QueryClient,
   QueryClientProvider,
   useMutation,
+  useQuery,
 } from '@tanstack/react-query';
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { getUntypedClient } from '@trpc/client';
-import { createTRPCReact } from '@trpc/react-query';
-import type { InferQueryLikeData } from '@trpc/react-query/shared';
+import {
+  createTRPCContext,
+  TRPCOptionsProxy,
+} from '@trpc/tanstack-react-query';
+// import type { InferQueryLikeData } from '@trpc/react-query/shared';
 import { konn } from 'konn';
 import type { ReactNode } from 'react';
 import React, { useState } from 'react';
@@ -106,25 +110,24 @@ describe('polymorphism', () => {
       } = createTRPCApi();
 
       const opts = routerToServerAndClientNew(appRouter);
-      const trpc = createTRPCReact<typeof appRouter>();
+
+      const { TRPCProvider, useTRPC } = createTRPCContext<typeof appRouter>();
 
       const queryClient = new QueryClient();
 
       function App(props: { children: ReactNode }) {
         return (
-          <trpc.Provider
-            {...{ queryClient, client: getUntypedClient(opts.client) }}
-          >
+          <TRPCProvider queryClient={queryClient} trpcClient={opts.client}>
             <QueryClientProvider client={queryClient}>
               {props.children}
             </QueryClientProvider>
-          </trpc.Provider>
+          </TRPCProvider>
         );
       }
       return {
         ...opts,
         App,
-        trpc,
+        useTRPC,
         IssueExportsProvider,
         DiscussionExportsProvider,
         PullRequestExportsProvider,
@@ -137,25 +140,26 @@ describe('polymorphism', () => {
 
   describe('simple factory', () => {
     test('can use a simple factory router with an abstract interface', async () => {
-      const { trpc } = ctx;
+      const { useTRPC } = ctx;
 
       /**
        * Can now define page components which re-use functionality from components,
        * and pass the specific backend functionality which is needed need
        */
       function IssuesExportPage() {
-        const utils = trpc.useUtils();
+        const trpc = useTRPC();
+        // const utils = trpc.useUtils();
 
         const [currentExport, setCurrentExport] = useState<number | null>(null);
 
         const invalidate = useMutation({
-          mutationFn: () => utils.invalidate(),
+          // mutationFn: () => utils.invalidate(),
         });
         return (
           <>
             <StartExportButton
               route={trpc.github.issues.export}
-              utils={utils.github.issues.export}
+              // utils={utils.github.issues.export}
               onExportStarted={setCurrentExport}
             />
 
@@ -202,21 +206,22 @@ describe('polymorphism', () => {
     });
 
     test('can use the abstract interface with a factory instance which has been merged with some extra procedures', async () => {
-      const { trpc } = ctx;
+      const { useTRPC } = ctx;
 
       function DiscussionsExportPage() {
-        const utils = trpc.useUtils();
+        const trpc = useTRPC();
+        // const utils = trpc.useUtils();
 
         const [currentExport, setCurrentExport] = useState<number | null>(null);
         const invalidate = useMutation({
-          mutationFn: () => utils.invalidate(),
+          // mutationFn: () => utils.invalidate(),
         });
 
         return (
           <>
             <StartExportButton
               route={trpc.github.discussions.export}
-              utils={utils.github.discussions.export}
+              // utils={utils.github.discussions.export}
               onExportStarted={setCurrentExport}
             />
 
@@ -265,18 +270,19 @@ describe('polymorphism', () => {
 
   describe('sub-typed factory', () => {
     test('can use a sub-typed factory router with the interfaces from the supertype', async () => {
-      const { trpc } = ctx;
+      const { useTRPC } = ctx;
 
       /**
        * Can define page components which operate over interfaces generated by a super-typed router,
        * but also extend types and functionality
        */
       function PullRequestsExportPage() {
-        const utils = trpc.useUtils();
+        const trpc = useTRPC();
+        // const utils = trpc.useUtils();
 
         const [currentExport, setCurrentExport] = useState<number | null>(null);
         const invalidate = useMutation({
-          mutationFn: () => utils.invalidate(),
+          // mutationFn: () => utils.invalidate(),
         });
 
         return (
@@ -284,7 +290,7 @@ describe('polymorphism', () => {
             {/* Some components may still need to be bespoke... */}
             <SubTypedStartExportButton
               route={trpc.github.pullRequests.export}
-              utils={utils.github.pullRequests.export}
+              // utils={utils.github.pullRequests.export}
               onExportStarted={setCurrentExport}
             />
 
@@ -295,7 +301,7 @@ describe('polymorphism', () => {
 
             <RemoveExportButton
               remove={trpc.github.pullRequests.export.delete}
-              utils={utils.github.pullRequests.export}
+              // utils={utils.github.pullRequests.export}
               exportId={currentExport}
             />
 
@@ -352,16 +358,18 @@ describe('polymorphism', () => {
 
 function StartExportButton(props: {
   route: Factory.ExportRouteLike;
-  utils: Factory.ExportUtilsLike;
+  // utils: Factory.ExportUtilsLike;
   onExportStarted: (id: number) => void;
 }) {
-  const exportStarter = props.route.start.useMutation({
-    async onSuccess(data) {
-      props.onExportStarted(data.id);
+  const exportStarter = useMutation(
+    props.route.start.mutationOptions({
+      async onSuccess(data) {
+        props.onExportStarted(data.id);
 
-      await props.utils.invalidate();
-    },
-  });
+        // await props.utils.invalidate();
+      },
+    }),
+  );
 
   return (
     <button
@@ -381,13 +389,15 @@ function StartExportButton(props: {
 function RemoveExportButton(props: {
   exportId: number | null;
   remove: SubTypedFactory.ExportSubTypedRouteLike['delete'];
-  utils: SubTypedFactory.ExportSubTypesUtilsLike;
+  // utils: SubTypedFactory.ExportSubTypesUtilsLike;
 }) {
-  const exportDeleter = props.remove.useMutation({
-    async onSuccess() {
-      await props.utils.invalidate();
-    },
-  });
+  const exportDeleter = useMutation(
+    props.remove.mutationOptions({
+      async onSuccess() {
+        // await props.utils.invalidate();
+      },
+    }),
+  );
 
   const id = props.exportId;
   if (!id) {
@@ -410,17 +420,19 @@ function RemoveExportButton(props: {
 
 type SubTypedStartExportButtonProps = {
   route: SubTypedFactory.ExportSubTypedRouteLike;
-  utils: SubTypedFactory.ExportSubTypesUtilsLike;
+  // utils: SubTypedFactory.ExportSubTypesUtilsLike;
   onExportStarted: (id: number) => void;
 };
 function SubTypedStartExportButton(props: SubTypedStartExportButtonProps) {
-  const exportStarter = props.route.start.useMutation({
-    onSuccess(data) {
-      props.onExportStarted(data.id);
+  const exportStarter = useMutation(
+    props.route.start.mutationOptions({
+      onSuccess(data) {
+        props.onExportStarted(data.id);
 
-      props.utils.invalidate();
-    },
-  });
+        // props.utils.invalidate();
+      },
+    }),
+  );
 
   return (
     <button
@@ -457,12 +469,14 @@ function ExportStatus<
   TStatus extends Factory.ExportRouteLike['status'],
 >(props: {
   status: TStatus;
-  renderAdditionalFields?: (data: InferQueryLikeData<TStatus>) => ReactNode;
+  renderAdditionalFields?: (data: TStatus['_output']) => ReactNode;
   currentExport: number | null;
 }) {
-  const exportStatus = props.status.useQuery(
-    { id: props.currentExport ?? -1 },
-    { enabled: props.currentExport !== null },
+  const exportStatus = useQuery(
+    props.status.queryOptions(
+      { id: props.currentExport ?? -1 },
+      { enabled: props.currentExport !== null },
+    ),
   );
 
   if (!exportStatus.data) {
@@ -473,13 +487,13 @@ function ExportStatus<
     <p>
       Last Export: `{exportStatus.data?.name}` (
       {exportStatus.data.downloadUri ? 'Ready!' : 'Working'})
-      {props.renderAdditionalFields?.(exportStatus.data as unknown as any)}
+      {props.renderAdditionalFields?.(exportStatus.data)}
     </p>
   );
 }
 
 function ExportsList(props: { list: Factory.ExportRouteLike['list'] }) {
-  const exportsList = props.list.useQuery();
+  const exportsList = useQuery(props.list.queryOptions());
 
   return (
     <>
